Swapped patients component table with dashboard component table
