+++ neutrackClientApp/src/app/nutritionist/dashboard/dashboard.component.html
@@ -1,54 +1,45 @@
 <div class="dashboardComponent">
 
     <div class="dashboardContent">
-
-        <h2>List of active patients</h2>
-        <mat-form-field>
-            <!-- <mat-label>Filter: </mat-label> -->
-            <input matInput (keyup)="applyFilter($event)" placeholder="Search by Name" #input>
-        </mat-form-field>
-
-        <div class="mat-elevation-z8">
-            <table mat-table [dataSource]="dataSource" matSort>
-        
-            <ng-container matColumnDef="id">
-                <th mat-header-cell *matHeaderCellDef mat-sort-header> ID </th>
-                <td mat-cell *matCellDef="let row"> {{row.id}} </td>
-            </ng-container>
-
-            <ng-container matColumnDef="name">
-                <th mat-header-cell *matHeaderCellDef mat-sort-header> Name </th>
-                <td mat-cell *matCellDef="let row"> {{row.name}} </td>
-            </ng-container>
-        
-            <ng-container matColumnDef="progress">
-                <th mat-header-cell *matHeaderCellDef mat-sort-header> DOB </th>
-                <td mat-cell *matCellDef="let row"> {{row.progress}}% </td>
-            </ng-container>
-
-            <ng-container matColumnDef="fruit">
-                <th mat-header-cell *matHeaderCellDef mat-sort-header> Start Date </th>
-                <td mat-cell *matCellDef="let row"> {{row.fruit}} </td>
-            </ng-container>
-        
-            <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-            <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-        
-            <!-- Row shown when there is no matching data. -->
-            <tr class="mat-row" *matNoDataRow>
-                <td class="mat-cell" colspan="4">No data matching the filter "{{input.value}}"</td>
-            </tr>
-            </table>
-        
-            <mat-paginator 
-              [length]="100"
-              [pageSize]="5"
-              [pageSizeOptions]="[5, 10, 25, 100]">
-            </mat-paginator>
-        </div>
+        <h2>List of the active patients</h2>
     </div>
 
     
-      
-</div>
 
+    <table mat-table
+            [dataSource]="dataSource" multiTemplateDataRows
+            class="mat-elevation-z8">
+        <ng-container matColumnDef="{{column}}" *ngFor="let column of columnsToDisplay">
+            <th mat-header-cell *matHeaderCellDef> {{column}} </th>
+            <td mat-cell *matCellDef="let element"> {{element[column]}} </td>
+        </ng-container>
+
+        <!-- Expanded Content Column - The detail row is made up of this one column that spans across all columns -->
+        <ng-container matColumnDef="expandedDetail">
+            <td mat-cell *matCellDef="let element" [attr.colspan]="columnsToDisplay.length">
+                <div class="example-element-detail"
+                    [@detailExpand]="element == expandedElement ? 'expanded' : 'collapsed'">
+                    <div class="patient-picture">
+                        <!-- <div class="example-element-position"> {{element.photo}} </div>       -->
+                    </div>
+                    <div class="patient-info">
+                        <div class="patient-current-weight"> Current weight: {{element.current_weight}} lb </div>
+                        <div class="patient-initial-weight"> Initial weight: {{element.weight}} lb </div>
+                        <div class="patient-weight"> Goal weight: {{element.weight}} lb </div>
+                        <div class="patient-height"> Height: {{element.height}} in</div>
+                        <!-- {{element.description}} -->
+                        <!-- <span class="example-element-description-attribution"> -- Wikipedia </span> -->
+                    </div>
+                </div>
+            </td>
+        </ng-container>
+
+        <tr mat-header-row *matHeaderRowDef="columnsToDisplay"></tr>
+        <tr mat-row *matRowDef="let element; columns: columnsToDisplay;"
+            class="example-element-row"
+            [class.example-expanded-row]="expandedElement === element"
+            (click)="expandedElement = expandedElement === element ? null : element">
+        </tr>
+        <tr mat-row *matRowDef="let row; columns: ['expandedDetail']" class="example-detail-row"></tr>
+    </table>
+</div>
