+++ neutrackClientApp/src/app/nutritionist/patients/patients.component.ts
@@ -1,67 +1,69 @@
-import { Component, OnInit } from '@angular/core';
-import {animate, state, style, transition, trigger} from '@angular/animations';
+import {AfterViewInit, Component, ViewChild} from '@angular/core';
+import {MatPaginator} from '@angular/material/paginator';
+import {MatSort} from '@angular/material/sort';
+import {MatTableDataSource} from '@angular/material/table';
+
+export interface UserData {
+  id: string;
+  name: string;
+  progress: string;
+  fruit: string;
+}
+
+/** Constants used to fill up our data base. */
+const FRUITS: string[] = [
+  'blueberry', 'lychee', 'kiwi', 'mango', 'peach', 'lime', 'pomegranate', 'pineapple'
+];
+const NAMES: string[] = [
+  'Maia', 'Asher', 'Olivia', 'Atticus', 'Amelia', 'Jack', 'Charlotte', 'Theodore', 'Isla', 'Oliver',
+  'Isabella', 'Jasper', 'Cora', 'Levi', 'Violet', 'Arthur', 'Mia', 'Thomas', 'Elizabeth'
+];
 
 @Component({
   selector: 'app-patients',
   templateUrl: './patients.component.html',
-  styleUrls: ['./patients.component.css'],
-  animations: [
-    trigger('detailExpand', [
-      state('collapsed', style({height: '0px', minHeight: '0'})),
-      state('expanded', style({height: '*'})),
-      transition('expanded <=> collapsed', animate('225ms cubic-bezier(0.4, 0.0, 0.2, 1)')),
-    ]),
-  ],
+  styleUrls: ['./patients.component.css']
 })
-export class PatientsComponent {
-  dataSource = ELEMENT_DATA;
-  // columnsToDisplay = ['name', 'weight', 'symbol', 'position'];
-  columnsToDisplay = ['Name', 'DOB', 'gender', 'email'];
-  expandedElement: PatientsList | null;
 
-  // constructor() { }
+export class PatientsComponent implements AfterViewInit {
+  displayedColumns: string[] = ['id', 'name', 'progress', 'fruit'];
+  dataSource: MatTableDataSource<UserData>;
+
+  @ViewChild(MatPaginator) paginator: MatPaginator;
+  @ViewChild(MatSort) sort: MatSort;
+
+  constructor() {
+    // Create 100 users
+    const users = Array.from({length: 100}, (_, k) => createNewUser(k + 1));
+
+    // Assign the data to the data source for the table to render
+    this.dataSource = new MatTableDataSource(users);
+  }
+
+  ngAfterViewInit() {
+    this.dataSource.paginator = this.paginator;
+    this.dataSource.sort = this.sort;
+  }
+
+  applyFilter(event: Event) {
+    const filterValue = (event.target as HTMLInputElement).value;
+    this.dataSource.filter = filterValue.trim().toLowerCase();
+
+    if (this.dataSource.paginator) {
+      this.dataSource.paginator.firstPage();
+    }
+  }
 }
 
-export interface PatientsList {
-  Name: string;
-  DOB: string;
-  gender: string;
-  email: string;
-  current_weight: number;
-  initial_weight: number;
-  goal_weight: number;
-  height: number;
-  
-}
+/** Builds and returns a new User. */
+function createNewUser(id: number): UserData {
+  const name = NAMES[Math.round(Math.random() * (NAMES.length - 1))] + ' ' +
+    NAMES[Math.round(Math.random() * (NAMES.length - 1))].charAt(0) + '.';
 
-const ELEMENT_DATA: PatientsList[] = [
-  {
-    
-    Name: 'Lilit Karayan',
-    DOB: '01/24/1992',
-    gender: 'F',
-    email: 'lkaraya1@my.westga.edu',
-    current_weight: 10,
-    initial_weight: 8,
-    goal_weight: 78,
-    height: 66
-  }, {
-    Name: 'Vincent Adeyemi',
-    DOB: '01/24/1992',
-    gender: 'F',
-    email: 'vadeyem1@my.westga.edu',
-    current_weight: 67,
-    initial_weight: 89,
-    goal_weight: 60,
-    height: 69
-  }, {
-    Name: 'Jordan Barron',
-    DOB: '01/24/1992',
-    gender: 'F',
-    email: 'barron8@my.westga.edu',
-    current_weight: 89,
-    initial_weight: 120,
-    goal_weight: 85,
-    height: 70
-  },  
-];
+  return {
+    id: id.toString(),
+    name: name,
+    progress: Math.round(Math.random() * 100).toString(),
+    fruit: FRUITS[Math.round(Math.random() * (FRUITS.length - 1))]
+  };
+}
